fix: unity build not working
