+++ packages/entryPoints/index.ts
@@ -1,42 +1,24 @@
 import 'engine'
-import * as qs from 'query-string'
 
 import { ETHEREUM_NETWORK, DEBUG } from '../config'
 import { initBabylonClient } from '../dcl'
-import { domReadyFuture, bodyReadyFuture } from '../engine/renderer/init'
+import { domReadyFuture, bodyReadyFuture, scene } from '../engine/renderer/init'
 import { initShared } from '../shared'
 import { enableParcelSceneLoading } from '../shared/world/parcelSceneManager'
+import { getWorldSpawnpoint } from '../shared/world/positionThings'
 import { WebGLParcelScene } from '../dcl/WebGLParcelScene'
 import { enableMiniMap } from '../dcl/widgets/minimap'
-import { movePlayerToSpawnpoint, teleportObservable } from '../shared/world/positionThings'
 
 export async function loadClient(net: ETHEREUM_NETWORK) {
   await initBabylonClient()
   document.body.appendChild(enableMiniMap())
 
-  let initialized = false
-  let spawnpointLand = qs.parse(location.search).position
-
-  teleportObservable.add(position => {
-    initialized = false
-    spawnpointLand = `${position.x},${position.y}`
-  })
-
   await enableParcelSceneLoading(net, {
     parcelSceneClass: WebGLParcelScene,
-    onLoadParcelScenes: lands => {
-      if (initialized) {
-        return
-      }
-
-      const land = lands.find(land => land.scene.scene.base === spawnpointLand)
-
-      if (!land) {
-        return
-      }
-
-      movePlayerToSpawnpoint(land)
-      initialized = true
+    onSpawnpoint: initialLand => {
+      const newPosition = getWorldSpawnpoint(initialLand)
+      const result = new BABYLON.Vector3(newPosition.x, newPosition.y, newPosition.z)
+      scene.activeCamera.position.copyFrom(result)
     },
     shouldLoadParcelScene: () => true
   })
